Adjusting some files to make the gem executable
