+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## [Unreleased]
+
+## [0.1.0] - 2024-04-04
+
+- Initial release
