+++ Gemfile
@@ -1,4 +1,9 @@
-source 'https://rubygems.org'
+# frozen_string_literal: true
 
-# Specify your gem's dependencies
+source "https://rubygems.org"
+
+# Specify your gem's dependencies in ollama.gemspec
 gemspec
+
+gem "rake", "~> 13.0"
+gem "rspec", "~> 3.0"
