+++ lib/ollama/ollama_generate_schema.rb
