+++ lib/ollama/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Ollama
+  VERSION = "0.1.0"
+end
