+++ sig/ollama.rbs
@@ -1,0 +1,4 @@
+module Ollama
+  VERSION: String
+  # See the writing guide of rbs: https://github.com/ruby/rbs#guides
+end
